requirements.txt
pandas>=2.2.0
pyyaml>=6.0
python-dotenv>=1.0.0
numpy>=1.26.0
pytest>=8.0.0
streamlit>=1.32.0
plotly>=5.19.0
